small fixes and improvements

+++ AutoSchedulerBE/AutoScheduler.Application/Services/TimesheetService.cs
@@ -48,6 +48,8 @@ namespace AutoScheduler.Application.Services
             int breakDuration,
             TimeOnly startTime,
             TimeOnly endTime,
+            TimeOnly? generalBreakStart,
+            TimeOnly? generalBreakEnd,
             ActivityRequirementsDTO[] requirements)
         {
             var result = mapper.MapResult(generatorOutput);
@@ -65,6 +67,8 @@ namespace AutoScheduler.Application.Services
                     BreakDuration = breakDuration,
                     StartTime = startTime,
                     EndTime = endTime,
+                    GeneralBreakStart = generalBreakStart,
+                    GeneralBreakEnd = generalBreakEnd,
                     Requirements = requirements,
                     Timeslots = timeslots
                 };
@@ -97,7 +101,14 @@ namespace AutoScheduler.Application.Services
             timesheetGenerator.Generate();
             var generatorOutput = timesheetGenerator.Generated;
 
-            return await TimesheetsFromGeneratorOutput(generatorOutput, mapper, generatorRequirementsDTO.SlotDurationInMinutes, generatorRequirementsDTO.BreakDurationInMinutes, generatorRequirementsDTO.StartTime, generatorRequirementsDTO.EndTime, generatorRequirementsDTO.Requirements);
+            return await TimesheetsFromGeneratorOutput(generatorOutput, mapper,
+                generatorRequirementsDTO.SlotDurationInMinutes,
+                generatorRequirementsDTO.BreakDurationInMinutes,
+                generatorRequirementsDTO.StartTime,
+                generatorRequirementsDTO.EndTime,
+                generatorRequirementsDTO.GeneralBreakStartTime,
+                generatorRequirementsDTO.GeneralBreakEndTime,
+                generatorRequirementsDTO.Requirements);
         }
 
         public async Task<IList<WeekDayTimeRangeDTO>> GetAvailableSpaceForTimeslotAsync(TimeslotPlacementChangeDTO timeslotPlacementChangeDTO)
@@ -304,6 +315,8 @@ namespace AutoScheduler.Application.Services
                 timeslotRearrangementDto.GeneratorRequirements.BreakDurationInMinutes,
                 timeslotRearrangementDto.GeneratorRequirements.StartTime,
                 timeslotRearrangementDto.GeneratorRequirements.EndTime,
+                timeslotRearrangementDto.GeneratorRequirements.GeneralBreakStartTime,
+                timeslotRearrangementDto.GeneratorRequirements.GeneralBreakEndTime,
                 timeslotRearrangementDto.GeneratorRequirements.Requirements);
         }
 
@@ -318,7 +331,7 @@ namespace AutoScheduler.Application.Services
 
         public async Task ActivateTimesheetAsync(int timesheetId)
         {
-            var timesheet = await _timesheetRepository.GetTimesheetByIdAsync(timesheetId);
+            var timesheet = await _timesheetRepository.GetTimesheetForUpdateAsync(timesheetId);
 
             var rootGroupIds = timesheet.Timeslots.Where(ts => !timesheet.Timeslots.Any(ts1 => ts.Group.ParentGroupId == ts1.GroupId)).Select(ts => ts.GroupId).Distinct();
             //disallow multiple active timesheets for (main) group
@@ -349,12 +362,15 @@ namespace AutoScheduler.Application.Services
 
         public async Task DeactivateTimesheetAsync(int timesheetId)
         {
-            var timesheetToDeactivate = await _timesheetRepository.GetTimesheetByIdAsync(timesheetId);
+            var timesheetToDeactivate = await _timesheetRepository.GetTimesheetForUpdateAsync(timesheetId);
+
+            if (timesheetToDeactivate == null || timesheetToDeactivate.Timeslots.IsNullOrEmpty())
+                throw new InvalidOperationException();
+            //delete availability entries corresponding to timeslots
+            await _timesheetRepository.DeleteTimeslotsAvailability(timesheetToDeactivate.Timeslots);
             //set state to inactive as a soft delete
             timesheetToDeactivate.State = TimesheetState.Inactive;
             await _timesheetRepository.UpdateTimesheetAsync(timesheetToDeactivate);
-            //delete availability entries corresponding to timeslots
-            await _timesheetRepository.DeleteTimeslotsAvailability(timesheetToDeactivate.Timeslots);
         }
 
         public async Task<IList<ActivityRequirementsDTO>> GetRequirementsForTimesheetAsync(int timesheetId)
